requirements.txt
# Framework web
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
email-validator>=2.1.0

# Sistema MongoDB
motor>=3.3.1
pymongo>=4.5.0
beanie>=1.21.0

# Segurança
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Processamento de dados
pandas>=2.1.1
numpy>=1.26.1
openpyxl>=3.1.2
xlrd>=2.0.1

# Ferramentas de validação
pydantic[email]>=2.4.2

# Testes
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0 